feat: add css styling to the whole site
As an assignment from lesson 'The Cascade', basic CSS styling was added
to the recipes page.

The font size was changed, and it now sizes more accurately to the
viewport of the device in which text is being displayed. Web-safe fonts
were included to provide a more secure font coverage on all devices.
Images were also resized according to the viewport's size.

A few color adjustments were made on the main contents, and different
parts of the recipes were color-coded, as well as center-aligned.