feat: add new styling options to homepage
These changes follow the TOP exercise at the end of the first couple of
CSS lessons.

A border to the overall page has been implemented with a dashed line for
a parallel to the dashed patterns sewn in some picnic fabric, smh.
There's also a more carefully built title to try and highlight it from
the rest of the elements. The list of recipes has also been revamped to
be more relevant in the page, and actually look like the main contents
of the site.